requirements.txt
numpy
pytorch
torchvision
tqdm
h5py
Pillow
click
matplotlib
tb-nightly # until 1.14